Allow deserialization of object values in examples (#132)
The OpenAPI3 [spec](https://swagger.io/specification/#example-object)
allows examples properties to contain object types. The existing code
was failing on object types in example properties.

This PR changes the deserialization of `ExampleValue` so object types
are parsed to strings.

However, I could not determine how the `ExampleValue` type is included
in the final `typed` object. My assumption was the value would be part
of the final `typed` object and could be used in documentation
generation (thus the string instead of a parsed object) but I wasn't
able to find the usage. Also, wasn't quite able to figure out how to
write tests for this case.